+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Updated to govuk-frontend@3.9.0  
   See https://github.com/alphagov/govuk-frontend/releases/tag/v3.9.0 for full release notes
 
+- New notification banner component added
+
 #### Breaking changes
 
 ---